pyproject.toml
[project]
name = "base_sites"
version = "1.0.1a0"
description = "Applications de base pour la construction d'autres sites plus complets"
authors = [
    {name = "Antoine Louatron",email = "antoine@louatron.fr"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<4"
dependencies = [
    "django (>=5.0, <6)",
    "redis[hiredis] (>=6.4.0,<7.0.0)",
    "bleach (>=6.2.0,<7.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "django-webpack-loader (>=3.2.1,<4.0.0)",
    "icalendar (>=6.3.1,<7.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "unidecode (>=1.4.0,<2.0.0)",
    "django-sendfile2 (>=0.7.2,<0.8.0)",
    "beautifulsoup4 (>=4.14.2,<5.0.0)",
    "wrapt (>=1.17.3,<2.0.0)",
    "lxml (>=6.0.2,<7.0.0)"
]

[tool.poetry]
package-mode = true
packages = [
    {include = "*", from="src"},
]

[tool.poetry.group.dev.dependencies]
coverage = "^7.10.7"
tox = "^4.30.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
